Change some wording and commas
I'm not sure if "level of Danish"/"niveau af dansk" sounds right either. But I think it works. I changed "tag" to "role" instead because, in my experience, people call them that or ranks. I hear tag too but then it'd be in neuter in Danish: "dit rette tag". Please tell if you disagree with any of my changes.